fix(network): harden SOCKS5 proxy replies
Avoid host DNS resolution of proxy-controlled reply addresses. CONNECT replies now parse domain-form bound addresses without resolving them, while UDP associations reject domain relay endpoints that cannot be used safely.\n\nMove proxy models and parsing out of the module root, and document that the UDP receive buffer already covers the complete outer datagram including its SOCKS5 header.
